requirements.txt
timm
pytorch
tqdm
scikit-learn
numpy
click